% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Genepop2SPAM.GCL}
\alias{Genepop2SPAM.GCL}
\title{Converts Genepop files to SPAM baseline and mixture files - IN DEVELOPMENT}
\usage{
Genepop2SPAM.GCL(baselineFile, mixtureFiles, baselineOutPath, mixtureOutPath)
}
\arguments{
\item{baselineFile}{The fullpath to the baseline Genepop file for the baseline in SPAM (.mix).}

\item{mixtureFiles}{A vector with the fullpath of the Genepop files to be created into seperate SPAM mixture (.mix) files.}

\item{baselineOutPath}{The full path to write the baseline file to}

\item{mixtureOutPath}{The full path to write all of the mixture files to.}
}
\description{
Genepop files contain the information to write the baseline ".bse" and mixture ".mix" files for SPAM. Use the following considerations when using this function.
}
\details{
\itemize{
  \item A. This function is set up to work only with diploid alleles for SNPs and microsatellites or mtDNA haplotypes with nine or less combined SNPs.
  \item B. All the files for the baseline and mixtures should be read in at the same time to create SPAM files.  This is because for each locus it will need to know how many different alleles are present across all files to properly id each one into the correct position.
  \item C. Each population in the Genepop file should be separated by POP, Pop, or pop.  Any other population separator will not be recognized.
  \item D. The locus names listed at the top of each Genepop file should be in the same order and have the same number of loci.
   ...
}
}
\keyword{ADFG}

